Update Sudoku.py
So far this has been the most success method.

Augmented a lot.

Changed how the check array operates. It now holds the values that don't violate the rules. If the list is has only one possibility then it will be assigned.

Currently there is more than one possibility for a couple, so the software can't continue.

Need to figure out a way to have it input values at this point.