Update PyGameStuff.py
Changed background colour for preset squares.

Made it so that you can't change these preset squares.

Fixed a bug that resulted from someone clicking on a  preset square in the reset/solve coluns.
This happened because of the way the conditional statements were set up.
Changing the else statement to an elif, renedued the bug.